Only read from cache if the filter matches exactly
1. Unless the filter matches exactly, it may be using results that are unsuitable for the filter. The remote data could be completely different given a different filter.
2. In said case, fetching from cache prevents it from requesting remote data that may be more suitable to the filter.

In any case, the old cache-pulling system is busted. It would match the 'a' in 'apple', cache the results, then later pull those same results for the 'ap' of apple. Because the results for 'a' and 'ap' were completely different in my case, none of the results were relevant and the autocomplete window wouldn't show.